ntu/Team#5-Unet_ImageSeg/unet_model/requirements.txt
numpy
matplotlib
argparse
Pillow
scipy
torch
torchaudio
torchvision
torchsummary
tqdm
pathlib